requirements.txt
streamlit
langchain_google_genai
google-generativeai
python-dotenv
PyPDF2
faiss-cpu
langchain